+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added `structures` in World Whitelist Object
+  * A string list of structures where the properties are applied
 * Time Existed, Condition and Depth modifiers now use a ModifiableValue instead of a simple number
 * Changed `override` to `keep_spawned` (so it now defaults to opposite behaviour)
 * Fixed crash when missing `biomes` or `dimensions` in world whitelist object